Changing error var names, moving errorMessage.ts,
