+++ lib/msal-core/src/AuthorityFactory.ts
@@ -28,7 +28,7 @@ import { Utils } from "./Utils";
 import { AadAuthority } from "./AadAuthority";
 import { B2cAuthority } from "./B2cAuthority";
 import { Authority, AuthorityType } from "./Authority";
-import { ErrorMessage } from "./ErrorMessage";
+import { ErrorMessage } from "./error/ErrorMessage";
 
 export class AuthorityFactory {
     /**
